requirements.txt
appnope==0.1.0
aspy.yaml==1.0.0
bleach==2.1.2
cached-property==1.3.1
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
google-api-core==0.1.2
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-translate==1.3.0
googleapis-common-protos==1.5.3
html5lib==1.0.1
identify==1.0.7
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nodeenv==1.2.0
notebook==5.2.2
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
pip-autoremove==0.9.0
pluggy==0.6.0
pre-commit==1.4.1
progressbar2==3.34.3
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
pyocr==0.4.7
pyparsing==2.2.0
PyQt5==5.9.2
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
scikit-image==0.13.1
scipy==1.0.0
simplegeneric==0.8.1
sip==4.19.6
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
tox==2.9.1
traitlets==4.3.2
translate==3.5.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.8
